move styles to themes folder
