--- app.780d9e983a479971d2de.css
@@ -1,1 +1,0 @@
-.editing{width:100%;height:10rem;resize:none}.comment{color:#999}.ns{margin:1rem;border:1px solid #ccc}.title{font-weight:700;font-size:1.5rem;margin-bottom:.5rem;border-bottom:1px dotted #ccc}.content,.title{padding:.5rem}.header{margin:1rem 0}.header a{float:right}.blogs{margin:1rem 0;border:1px solid #ccc}.ns{padding:1rem}.ns h1{float:left}.ns ul{float:right}.ns li{float:left;margin:0 0 0 1rem}.ns a{font-size:1rem;line-height:1rem}.ns{position:absolute;width:100%;background:#000;box-sizing:border-box;padding:1rem;height:3rem}.main{padding:4rem 1rem 1rem}
